feat: migrate dev pages to root
