+++ apps/web/src/app/logout/page.tsx
@@ -16,7 +16,7 @@ export default function LogoutPage() {
         // ignore errors
       }
       auth.logout();
-      router.replace('/dev');
+      router.replace('/');
     })();
   }, [router]);
