+++ apps/web/src/components/header.tsx
@@ -26,44 +26,31 @@ export default function Header({ bgColor, textColor }: HeaderProps) {
 
   return (
     <header style={{ backgroundColor: bgColor }}>
-      <nav className="container mx-auto flex items-center justify-between p-4">
-        <ul className="flex gap-4">
+      <nav className="container mx-auto flex items-center p-4">
+        <Link href="/" className="text-xl font-bold" style={linkStyle}>
+          Afterlight
+        </Link>
+        <ul className="flex flex-1 justify-center gap-4">
           <li>
-            <Link
-              href="/dev"
-              className="flex items-center gap-1"
-              style={linkStyle}
-            >
+            <Link href="/" className="flex items-center gap-1" style={linkStyle}>
               <Home className="h-4 w-4" />
               <span>Главная</span>
             </Link>
           </li>
           <li>
-            <Link
-              href="/dev/how"
-              className="flex items-center gap-1"
-              style={linkStyle}
-            >
+            <Link href="/how" className="flex items-center gap-1" style={linkStyle}>
               <Info className="h-4 w-4" />
               <span>Как это работает</span>
             </Link>
           </li>
           <li>
-            <Link
-              href="/dev/policies"
-              className="flex items-center gap-1"
-              style={linkStyle}
-            >
+            <Link href="/policies" className="flex items-center gap-1" style={linkStyle}>
               <FileText className="h-4 w-4" />
               <span>Политики</span>
             </Link>
           </li>
           <li>
-            <Link
-              href="/dev/contacts"
-              className="flex items-center gap-1"
-              style={linkStyle}
-            >
+            <Link href="/contacts" className="flex items-center gap-1" style={linkStyle}>
               <Mail className="h-4 w-4" />
               <span>Контакты</span>
             </Link>
@@ -72,11 +59,7 @@ export default function Header({ bgColor, textColor }: HeaderProps) {
         <ul className="flex gap-4">
           {role === 'guest' ? (
             <li>
-              <Link
-                href="/dev/login"
-                className="flex items-center gap-1"
-                style={linkStyle}
-              >
+              <Link href="/login" className="flex items-center gap-1" style={linkStyle}>
                 <LogIn className="h-4 w-4" />
                 <span>Войти</span>
               </Link>
@@ -86,7 +69,7 @@ export default function Header({ bgColor, textColor }: HeaderProps) {
               {role === 'owner' && (
                 <li>
                   <Link
-                    href="/dev/owner"
+                    href="/owner"
                     className="flex items-center gap-1"
                     style={linkStyle}
                   >
@@ -98,7 +81,7 @@ export default function Header({ bgColor, textColor }: HeaderProps) {
               {role === 'verifier' && (
                 <li>
                   <Link
-                    href="/dev/verifier"
+                    href="/verifier"
                     className="flex items-center gap-1"
                     style={linkStyle}
                   >
@@ -109,7 +92,7 @@ export default function Header({ bgColor, textColor }: HeaderProps) {
               )}
               <li>
                 <Link
-                  href="/dev/logout"
+                  href="/logout"
                   className="flex items-center gap-1"
                   style={linkStyle}
                 >
