+++ docs/web.md
@@ -1,6 +1,6 @@
 # Frontend overview
 
-Веб-часть построена на Next.js. Корневая страница (`/`) служит временным лендингом и содержит ссылки на основные экраны приложения.
+Веб-часть построена на Next.js. Корневая страница (`/`) содержит основной интерфейс приложения.
 
 ## Навигация
 - `/wireframes/owner` — кабинет владельца (wireframe)
@@ -25,7 +25,7 @@
 - **Admin** — управляет сервисом через `/adm`.
 
 ## Авторизация ролей
-Страницы `/dev/owner` и `/dev/verifier` доступны только после проверки токена через `/auth/me`. При отсутствии cookie сессии пользователь перенаправляется на `/dev/login`. На клиенте компоненты дополнительно сверяют роль и, если она не соответствует странице, показывают сообщение «Доступ запрещён».
+Страницы `/owner` и `/verifier` доступны только после проверки токена через `/auth/me`. При отсутствии cookie сессии пользователь перенаправляется на `/login`. На клиенте компоненты дополнительно сверяют роль и, если она не соответствует странице, показывают сообщение «Доступ запрещён».
 
 ## Перезапуск seed
 Повторно заполнить БД можно так:
